Added three dialogs to story editor.
One for if all the ac's are removed and the story state therefore has to be changed to none. Two for if the estimate is changed to not estimated and this also affect readiness. Three for if the story state is changed directly. These all affect whether or not a story can be included in a sprint.

story:256